server/requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.23
sqlalchemy-serializer==1.4.1
openai==1.3.9
python-dotenv==1.0.0
Flask-CORS==4.0.0 